requirements.txt
fastapi
uvicorn
pydantic
groq
pypdf
python-dotenv